Adds the posibility to use the WPCOM theme in the Blaze Dashboad calypso app
