+++ apps/blaze-dashboard/src/components/generic-header/blaze-plugin-header/style.scss
@@ -1,7 +1,7 @@
 @import "@automattic/color-studio/dist/color-variables";
 @import "@automattic/components/src/styles/typography";
 
-.woo-blaze-header {
+.blaze-plugin-header {
 	display: flex;
 	flex-direction: column;
