+++ apps/blaze-dashboard/src/components/generic-header/jetpack-blaze-header/style.scss
