+++ apps/blaze-dashboard/src/load-config.js
@@ -12,6 +12,9 @@ if ( window.configData?.blog_id ) {
 const isWooStore = window.configData?.is_woo_store || false;
 const needSetup = window.configData?.need_setup || false;
 
+// Using the isWooStore fallback condition for compatibility: That flag only comes true if they are using the Blaze plugin
+const isBlazePlugin = window.configData?.is_blaze_plugin || isWooStore;
+
 // The JSON is filtered by `apps/blaze-dashboard/filter-json-config-loader.js`.
 import productionConfig from '../../../config/production.json';
 
@@ -19,7 +22,10 @@ import productionConfig from '../../../config/production.json';
 productionConfig.features.is_running_in_jetpack_site =
 	window.configData.features.is_running_in_jetpack_site ?? true;
 
-// Set is is_running_in_woo_site to true if the dashboard is running on the Woo Blaze plugin
+// Set is is_running_in_woo_site to true if the dashboard is running on the Blaze Ads plugin
+productionConfig.features.is_running_in_blaze_plugin = isBlazePlugin;
+
+// Set is is_running_in_woo_site to true if the dashboard is running on a Woo site
 productionConfig.features.is_running_in_woo_site = isWooStore;
 
 productionConfig.features.blaze_setup_mode = !! needSetup;
@@ -29,7 +35,7 @@ productionConfig.features.blaze_setup_mode = !! needSetup;
 productionConfig.features[ 'use-translation-chunks' ] = false;
 
 // Sets the advertising path prefix for this app
-productionConfig.advertising_dashboard_path_prefix = isWooStore ? '/wc-blaze' : '/advertising';
+productionConfig.advertising_dashboard_path_prefix = isBlazePlugin ? '/wc-blaze' : '/advertising';
 
 // Note: configData is hydrated in Jetpack: https://github.com/Automattic/jetpack/blob/60b0dac0dc5ad7edec2b86edb57b65a3a98ec42d/projects/packages/blaze/src/class-dashboard-config-data.php#L31
 window.configData = {
